requirements.txt
tensorflow 
keras 
numpy 
pandas 
pillow